services/ocr-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
paddleocr==2.8.1
paddlepaddle==2.6.2
